% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{penn.demo.all}
\alias{penn.demo.all}
\title{Generate demo plot for all palettes.}
\usage{
penn.demo.all()
}
\value{
A list of ggplot2 bar plots.
}
\description{
Generate demo plot for all palettes.
}
\examples{
penn.demo.all()

}
