% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{penn.demo}
\alias{penn.demo}
\title{Generate demo plot for single palette.}
\usage{
penn.demo(palette_name)
}
\arguments{
\item{palette_name}{The short ID for the palette, e.g. "Blue 8".}
}
\value{
A ggplot2 bar plot.
}
\description{
Generate demo plot for single palette.
}
\examples{
penn.demo("Red-Blue")

}
