% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{penn}
\alias{penn}
\title{Generate color palette.}
\usage{
penn(palette_name)
}
\arguments{
\item{palette_name}{The short ID for the palette, e.g. "Blue 8".}
}
\value{
A vector of hexademical values.
}
\description{
Generate color palette.
}
\examples{
# Assign palette to a name.
penn_palette <- penn("Red-Blue")

# Use with base R.
plot(iris$Sepal.Width,
     iris$Sepal.Length,
     col = penn("Blue 5"))

}
