% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{penn_palettes}
\alias{penn_palettes}
\title{List of all palettes.}
\format{
An object of class \code{list} of length 13.
}
\usage{
penn_palettes
}
\description{
List of all palettes.
}
\keyword{datasets}
